Rename util/class to util/bot-helpers.
