+++ src/util/bot-helpers.js
@@ -1,4 +1,4 @@
-
+// Copy whitelisted source properties to target object.
 export function overrideProperties(target, source, overrides = []) {
   overrides.forEach(name => {
     if (source[name]) {
